requirements.txt
pdf2image
google-generativeai
python-dotenv
tqdm
requests
pypdf
Pillow 